AI/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
torch==2.5.1
torchvision==0.20.1
Pillow==11.0.0
python-multipart==0.0.17
rembg==2.0.59